% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arena_class.R
\docType{data}
\name{Arena}
\alias{Arena}
\alias{coach}
\title{arena class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
coach
}
\description{
An Arena class where any 2 agents can be pit against each other.
}
\keyword{datasets}
