% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{CanonicalForm.othello}
\alias{CanonicalForm.othello}
\title{get canonical form of the othello class}
\usage{
\method{CanonicalForm}{othello}(game)
}
\arguments{
\item{game}{othello object}
}
\description{
get canonical form of the othello class
}
