% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation_use_boardgame.R
\name{chk_0_sp1}
\alias{chk_0_sp1}
\title{check whether the space is 0 and eligible}
\usage{
chk_0_sp1(df, search_color)
}
\arguments{
\item{df}{board map}

\item{search_color}{color of the opposite piece}
}
\description{
check whether the space is 0 and eligible
}
