% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation.R
\name{chk_eight_ds}
\alias{chk_eight_ds}
\title{check 8 directions around}
\usage{
chk_eight_ds(df, row, col)
}
\arguments{
\item{df}{board state}

\item{row}{row of the piece}

\item{col}{col of the piece}
}
\description{
check 8 directions around
}
