% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation.R
\name{chk_one_direction}
\alias{chk_one_direction}
\title{check directions}
\usage{
chk_one_direction(df, row, col, piece_color, search_color, x, y,
  place = T)
}
\arguments{
\item{df}{board}

\item{row}{row of the piece}

\item{col}{col of the piece}

\item{piece_color}{color to check for continuity}

\item{search_color}{color to search}

\item{x}{horizontal direction c(0, 1, -1)}

\item{y}{veritcal direction c(0, 1, -1)}

\item{place}{T for new piece to play, F for pieces to flip}
}
\description{
check directions
}
