% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation.R
\name{chk_valid_move_1d}
\alias{chk_valid_move_1d}
\title{check valid move one direction}
\usage{
chk_valid_move_1d(df, row, col, x, y)
}
\arguments{
\item{df}{board state}

\item{row}{row of the piece}

\item{col}{col of the piece}

\item{x}{horizontal direction}

\item{y}{vertical direction}
}
\description{
check valid move one direction
}
