% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation.R
\name{generate_othello_base}
\alias{generate_othello_base}
\title{generate the default board}
\usage{
generate_othello_base(sz)
}
\arguments{
\item{sz}{size of the board}
}
\description{
generate the default board
}
