% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation_use_boardgame.R
\name{generate_othello_base_M}
\alias{generate_othello_base_M}
\title{generate the default board matrix}
\usage{
generate_othello_base_M(sz, val = NULL)
}
\arguments{
\item{sz}{size of the board}

\item{val}{value of the board}
}
\description{
generate the default board matrix
}
