% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{get_othello_mv}
\alias{get_othello_mv}
\title{othello get moves using matrix}
\usage{
get_othello_mv(df, search_player = -1)
}
\arguments{
\item{df}{board state}

\item{search_player}{opponents color}
}
\description{
othello get moves using matrix
}
