% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation.R
\name{get_pieces_to_filp}
\alias{get_pieces_to_filp}
\alias{get_pieces_to_flip}
\title{get piecies to flip}
\usage{
get_pieces_to_flip(df, row, col, color)
}
\arguments{
\item{df}{board}

\item{row}{row of the new piece}

\item{col}{col of the new piece}

\item{color}{color of the new piece}
}
\description{
get piecies to flip
}
