% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS_zero.R
\name{getActionProb.MTCSzero}
\alias{getActionProb.MTCSzero}
\title{get action probability of MTCSzero}
\usage{
\method{getActionProb}{MTCSzero}(MTCSzero, canonicalBoard, temp = 1)
}
\arguments{
\item{MTCSzero}{MTCSzero object}

\item{canonicalBoard}{canonical Board position}
}
\description{
get action probability of MTCSzero
}
