% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS_zero.R
\name{getActionProb}
\alias{getActionProb}
\title{get action probability of MTCSzero}
\usage{
getActionProb(MTCSzero, canonicalBoard, temp = 1)
}
\arguments{
\item{MTCSzero}{MTCSzero object}

\item{canonicalBoard}{canonical Board position}

\item{args}{arguments}
}
\description{
get action probability of MTCSzero
}
