% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{getActionSize.othello}
\alias{getActionSize.othello}
\title{othello get action size}
\usage{
\method{getActionSize}{othello}(game)
}
\arguments{
\item{game}{othello object}
}
\description{
othello get action size
}
