% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{getBoardSize.othello}
\alias{getBoardSize.othello}
\title{othello get board size}
\usage{
\method{getBoardSize}{othello}(game)
}
\arguments{
\item{game}{othello object}
}
\description{
othello get board size
}
