% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{getGameEnded.othello}
\alias{getGameEnded.othello}
\title{othello check whether game has ended}
\usage{
\method{getGameEnded}{othello}(game)
}
\arguments{
\item{game}{othello object}
}
\description{
othello check whether game has ended
}
