% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{getInitBoard.othello}
\alias{getInitBoard.othello}
\title{othello get initial board}
\usage{
\method{getInitBoard}{othello}(game)
}
\arguments{
\item{game}{othello game object}
}
\description{
othello get initial board
}
