% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{getNextState.othello}
\alias{getNextState.othello}
\title{othello make moves using matrix}
\usage{
\method{getNextState}{othello}(game, move)
}
\arguments{
\item{game}{othello object}

\item{move}{move to make}
}
\description{
othello make moves using matrix
}
