% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{getSymmetries.othello}
\alias{getSymmetries.othello}
\title{get Symmetries of the othello class}
\usage{
\method{getSymmetries}{othello}(game, pi)
}
\arguments{
\item{game}{othello object}
}
\description{
get Symmetries of the othello class
}
