% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{getValidMove.othello}
\alias{getValidMove.othello}
\title{othello get moves using matrix}
\usage{
\method{getValidMove}{othello}(game)
}
\arguments{
\item{game}{othello object}
}
\description{
othello get moves using matrix
}
