% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet_models.R
\name{ini_nnet_othello}
\alias{ini_nnet_othello}
\title{create neural net for othello}
\usage{
ini_nnet_othello(game, args)
}
\arguments{
\item{game}{othello game state}

\item{args}{nnet arguments}
}
\description{
create neural net for othello
}
