% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{ini_othello}
\alias{ini_othello}
\title{othello state class}
\usage{
ini_othello(sz, val = NULL, player = 1)
}
\arguments{
\item{sz}{size of the board}

\item{val}{board values}
}
\description{
othello state class
}
