% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS_zero.R
\name{init_MTCSzero}
\alias{init_MTCSzero}
\title{initiate MTCS zero}
\usage{
init_MTCSzero(game, nnet, args = list(EPS = 0.00000001, numMCTSSims =
  50))
}
\arguments{
\item{game}{game object, for example result of ini_othello}

\item{nnet}{prediction neural network}

\item{args}{arguments, numMCTSSims number of MCTS simulations}
}
\description{
initiate MTCS zero
}
