% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS.R
\name{init_node}
\alias{init_node}
\alias{ini_node}
\title{node class definition}
\usage{
ini_node(MT = NULL, move = NA, parent = NA, state = NULL)
}
\arguments{
\item{MT}{the tree to initiate node into}

\item{move}{the move to make}

\item{parent}{id of parent}

\item{state}{game state}
}
\description{
node class definition
}
