% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS.R
\name{init_node1}
\alias{init_node1}
\alias{ini_node1}
\title{node class definition}
\usage{
ini_node1(MT = NULL, move = 0, parent = 0, parent_move = 0,
  state = NULL)
}
\arguments{
\item{MT}{the tree to initiate node into}

\item{move}{the move to make}

\item{parent}{id of parent}

\item{state}{game state}
}
\description{
node class definition
}
