% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation.R
\name{make_legal_move}
\alias{make_legal_move}
\title{make legal moves}
\usage{
make_legal_move(df, row, col, color)
}
\arguments{
\item{df}{board}

\item{legal_move_df}{legal move data frame}
}
\description{
make legal moves
}
