% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation_use_boardgame.R
\name{mk_a_move}
\alias{mk_a_move}
\title{make legal move}
\usage{
mk_a_move(df, legal_move_df, id, color)
}
\arguments{
\item{df}{board map}

\item{legal_move_df}{legal move space}

\item{id}{id of the legal move}

\item{color}{color of the legal move}
}
\description{
make legal move
}
