% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS_zero.R
\name{MTCSsearch.MTCSzero}
\alias{MTCSsearch.MTCSzero}
\title{perform one search of MTCS}
\usage{
\method{MTCSsearch}{MTCSzero}(MTCSzero, canonicalBoard)
}
\arguments{
\item{MTCSzero}{MTCSzero object}

\item{canonicalBoard}{canonical Board position}
}
\description{
perform one search of MTCS
}
