% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet_models.R
\docType{data}
\name{nnetwrapper}
\alias{nnetwrapper}
\title{nnet wrapper}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
nnetwrapper
}
\description{
nnet wrapper
}
\keyword{datasets}
