% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{othello_CanonicalForm}
\alias{othello_CanonicalForm}
\title{get canonical form of the othello class}
\usage{
othello_CanonicalForm(obj)
}
\arguments{
\item{obj}{othello object}
}
\description{
get canonical form of the othello class
}
