% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{othello_getGameEnded}
\alias{othello_getGameEnded}
\title{othello check whether game has ended}
\usage{
othello_getGameEnded(df, player)
}
\arguments{
\item{df}{othello board}

\item{player}{1 or -1}
}
\description{
othello check whether game has ended
}
