% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{othello_stringRepresentation}
\alias{othello_stringRepresentation}
\title{othello string representation}
\usage{
othello_stringRepresentation(obj)
}
\arguments{
\item{obj}{othello object}
}
\description{
othello string representation
}
