% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation.R
\name{play_randomly}
\alias{play_randomly}
\title{play randomly}
\usage{
play_randomly(df, color)
}
\arguments{
\item{df}{board}

\item{color}{color of the piece on turn}
}
\description{
play randomly
}
