% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_operation.R
\name{play_randomly1}
\alias{play_randomly1}
\title{play randomly function1}
\usage{
play_randomly1(df, color)
}
\arguments{
\item{df}{board}

\item{color}{color of the piece on turn}
}
\description{
play randomly function1
}
