% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS_zero.R
\name{search_MTCSzero}
\alias{search_MTCSzero}
\title{perform one search of MTCS}
\usage{
search_MTCSzero(MTCSzero, canonicalBoard)
}
\arguments{
\item{MTCSzero}{MTCSzero object}

\item{canonicalBoard}{canonical Board position}
}
\description{
perform one search of MTCS
}
