% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{search_neighbor}
\alias{search_neighbor}
\title{search moves using matrix}
\usage{
search_neighbor(mat, search_player = -1)
}
\arguments{
\item{mat}{board matrix}

\item{search_player}{opponents color}
}
\description{
search moves using matrix
}
