% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{stringRepresentation.othello}
\alias{stringRepresentation.othello}
\title{othello string representation}
\usage{
\method{stringRepresentation}{othello}(game)
}
\arguments{
\item{game}{othello object}
}
\description{
othello string representation
}
