% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{summary.othello_state}
\alias{summary.othello_state}
\title{summary method for othello game state}
\usage{
\method{summary}{othello_state}(x)
}
\arguments{
\item{x}{anything to be summaried}
}
\description{
summary method for othello game state
}
