% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS.R
\name{UCT}
\alias{UCT}
\title{UCT function}
\usage{
UCT(rootstate, itermax, verbose = FALSE)
}
\arguments{
\item{rootstate}{state of the root}

\item{itermax}{max iterations}

\item{verbose}{whether to print results}
}
\description{
UCT function
}
