% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS_zero.R
\name{UCTNN}
\alias{UCTNN}
\title{UCT with NNet function}
\usage{
UCTNN(MT = NULL, rootstate, NN, itermax, verbose = FALSE)
}
\arguments{
\item{MT}{monte carlo Tree}

\item{rootstate}{state of the root}

\item{NN}{neural net model}

\item{itermax}{max iterations}

\item{verbose}{whether to print results}
}
\description{
UCT with NNet function
}
