% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS.R
\name{UCTSelectChild}
\alias{UCTSelectChild}
\title{select child note}
\usage{
UCTSelectChild(node, UCTK = 1)
}
\arguments{
\item{node}{a mtcs node}

\item{UCTK}{constant to control exploration vs exploitation}
}
\description{
select child note
}
