% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCS.R
\name{UCTSelectChild1}
\alias{UCTSelectChild1}
\title{select child note}
\usage{
UCTSelectChild1(MTCS_value, node_Id, nodemove, UCTK = 1)
}
\arguments{
\item{MTCS_value}{MTCS_Value table}

\item{node_Id}{node Id}

\item{nodemove}{node move}

\item{UCTK}{constant to control exploration vs exploitation}
}
\description{
select child note
}
