% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/othello_class.R
\name{othello_Symmetries}
\alias{othello_Symmetries}
\title{get Symmetries of the othello class}
\usage{
othello_Symmetries(game)
}
\arguments{
\item{game}{othello object}
}
\description{
get Symmetries of the othello class
}
